refactor: move search code to site/search
