+++ baker/siteRenderers.tsx
@@ -7,7 +7,7 @@ import { BlogIndexPage } from "../site/BlogIndexPage.js"
 import { FrontPage } from "../site/FrontPage.js"
 import { ChartsIndexPage, ChartIndexItem } from "../site/ChartsIndexPage.js"
 import { CovidPage } from "../site/CovidPage.js"
-import { SearchPage } from "../site/SearchPage.js"
+import { SearchPage } from "../site/search/SearchPage.js"
 import { NotFoundPage } from "../site/NotFoundPage.js"
 import { DonatePage } from "../site/DonatePage.js"
 import OwidArticlePage from "../site/gdocs/OwidArticlePage.js"
