+++ site/search/SearchPageMain.tsx
@@ -2,7 +2,7 @@ import ReactDOM from "react-dom"
 import React from "react"
 import { getWindowQueryParams } from "@ourworldindata/utils"
 import { siteSearch, SiteSearchResults } from "./searchClient.js"
-import { SearchResults } from "../site/SearchResults.js"
+import { SearchResults } from "./SearchResults.js"
 import { observer } from "mobx-react"
 import { action, observable, runInAction } from "mobx"
